finish state machine and styling of <PersonalDetails /> & <PreviousOrders /> components
