+++ chowpronto_frontend/src/assets/svgs/settingsPage/index.js
@@ -1,0 +1,7 @@
+import OptionsSvg from "./options_svg.svg";
+
+const svg = {
+  src: OptionsSvg,
+  alt: "Menu options icon. Click to view details",
+};
+export default svg;
