+++ chowpronto_frontend/src/patron/pages/settings/components/styledComponents.js
@@ -1,0 +1,29 @@
+import styled from "styled-components";
+
+export const InnerTitle = styled.h2`
+  font-weight: normal;
+  font-size: 30px;
+  line-height: 35px;
+  text-align: left;
+  margin: 0;
+`;
+export const BorderedContainer = styled.div`
+  border: 1px solid #3949ab;
+  box-sizing: border-box;
+  border-radius: 10px;
+  padding: 2em 2.9em;
+`;
+export const Title = styled.h2`
+  color: #2f3249;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 21px;
+`;
+
+export const Grid = styled.div`
+  display: grid;
+  grid-template-columns: 0.4fr 1fr 100px;
+  cursor: pointer;
+  align-items: center;
+  height: 5em;
+`;
